MAGE-1298: update config tests + manage renderingConent exception
